// src/types/connection-options.ts
export interface ConnectionOptions {
  user?: string;
  pass?: string;
  port?: number;
  activePort?: number;
  activeIp?: string;
  activeIpv6?: boolean;
  secure?: boolean;
}

export interface IntConnOpts {
  user: string;
  pass: string;
  port: number;
  activePort: number;
  activeIp: string;
  activeIpv6: boolean;
  secure: boolean;
}